remove dmt database connection
