+++ api/core/rest/Actions.py
@@ -4,8 +4,7 @@ from flask import Blueprint, Response, request
 
 from classes.dto import DTO
 
-from classes.data_source import DataSource
-from core.repository.repository_factory import get_repository
+from core.repository.repository_factory import get_data_source
 
 
 class ApiActions(Enum):
@@ -24,10 +23,7 @@ def process_action():
 
 class ProcessAction:
     def __init__(self, request_data):
-        data_source_id = request_data["dataSource"]
-        data_source = DataSource(uid=data_source_id)
-        self.data_source = DataSource(uid=data_source_id)
-        self.document_repository = get_repository(data_source)
+        self.document_repository = get_data_source(request_data["dataSource"])
 
         # shared common request data properties
         self.action = request_data["action"]
