+++ api/core/rest/System.py
@@ -3,7 +3,7 @@ import json
 from flask import Blueprint, Response, send_file
 
 from core.enums import STATUS_CODES
-from core.repository.repository_factory import get_repository
+from core.repository.repository_factory import get_data_source
 from core.shared import request_object as req
 from core.shared import response_object as res
 from core.use_case.create_application_use_case import CreateApplicationRequestObject, CreateApplicationUseCase
@@ -18,7 +18,7 @@ blueprint = Blueprint("system", __name__)
 @blueprint.route("/api/v2/system/<string:data_source_id>/create-application/<string:application_id>", methods=["GET"])
 def post(data_source_id: str, application_id: str):
     logger.info(f"Creating application in data source '{data_source_id}' from application settings '{application_id}'")
-    document_repository = get_repository(data_source_id)
+    document_repository = get_data_source(data_source_id)
     request_object = CreateApplicationRequestObject.from_dict({"applicationId": application_id})
     use_case = CreateApplicationUseCase(document_repository)
     response = use_case.execute(request_object)
@@ -50,11 +50,11 @@ def get():
 @blueprint.route("/api/v2/system/<string:data_source_id>/generate-python-code/<string:document_id>", methods=["GET"])
 def generate_python_code(data_source_id: str, document_id: str):
     logger.info(f"Compiling the blueprint '{document_id}', in '{data_source_id}' to Python code")
-    document_repository = get_repository(data_source_id)
+    document_repository = get_data_source(data_source_id)
     request_object = GeneratePythonCodeRequestObject.from_dict(
         {"documentId": document_id, "dataSourceId": data_source_id}
     )
-    use_case = GeneratePythonCodeUseCase(document_repository, get_repository)
+    use_case = GeneratePythonCodeUseCase(document_repository, get_data_source)
     response = use_case.execute(request_object)
 
     if response.type == res.ResponseSuccess.SUCCESS:
