+++ api/core/use_case/get_document_by_path_use_case.py
@@ -1,5 +1,5 @@
 from core.enums import PRIMITIVES
-from core.repository.repository_factory import get_repository
+from core.repository.repository_factory import get_data_source
 from core.service.document_service import DocumentService
 from core.shared import request_object as req
 from core.shared import response_object as res
@@ -41,7 +41,7 @@ class GetDocumentByPathRequestObject(req.ValidRequestObject):
 
 
 class GetDMTDocumentByPathUseCase(uc.UseCase):
-    def __init__(self, repository_provider=get_repository):
+    def __init__(self, repository_provider=get_data_source):
         self.repository_provider = repository_provider
         self.document_service = DocumentService(repository_provider=self.repository_provider)
