+++ api/core/use_case/instantiate_entity.py
@@ -3,11 +3,11 @@ from core.shared import response_object as res
 from core.shared import use_case as uc
 from core.service.document_service import DocumentService
 
-from core.repository.repository_factory import get_repository
+from core.repository.repository_factory import get_data_source
 
 
 class InstantiateEntityUseCase(uc.UseCase):
-    def __init__(self, repository_provider=get_repository):
+    def __init__(self, repository_provider=get_data_source):
         self.repository_provider = repository_provider
 
     def process_request(self, request_object):
