+++ api/core/utility.py
@@ -8,7 +8,6 @@ from classes.tree_node import Node
 from config import Config
 from core.repository.repository_exceptions import EntityNotFoundException
 from services.data_modelling_document_service import document_api
-from services.database import dmt_database
 from utils.logging import logger
 
 
@@ -23,14 +22,6 @@ def url_safe_name(name: str) -> bool:
     match = expression.match(name)
     if match:
         return True
-
-
-def wipe_db():
-    print("Dropping all collections")
-    # FIXME: Read names from the database
-    for name in [Config.BLUEPRINT_COLLECTION, Config.ENTITY_COLLECTION, Config.SYSTEM_COLLECTION, "documents"]:
-        print(f"Dropping collection '{name}'")
-        dmt_database.drop_collection(name)
 
 
 # TODO: This should be removed when all code is using DMSS_API
